Working fileVersion component test w/factory
[ci skip]
